refactor(application): organize effect workflows by domain
