+++ application/media/__init__.py
@@ -1,17 +1,1 @@
-"""Media application use cases."""
-
-from .effects import (
-    EffectExportResult,
-    EffectOperation,
-    EffectRequest,
-    EffectUseCase,
-    validate_effect_storage_name,
-)
-
-__all__ = [
-    "EffectExportResult",
-    "EffectOperation",
-    "EffectRequest",
-    "EffectUseCase",
-    "validate_effect_storage_name",
-]
+"""Cross-domain media application capabilities."""
