+++ test/unit/application/effects/__init__.py
@@ -1,0 +1,1 @@
+"""Effect application tests."""
